backend/requirements.txt
fastapi
uvicorn
pillow
easyocr
opencv-python-headless
numpy
spacy
transformers
torch
anthropic
httpx
python-multipart
python-dotenv
sqlalchemy
passlib[bcrypt]
bcrypt
python-jose[cryptography]
fastapi-mail
pydantic[email]
requests
httpx
pytesseract
rapidfuzz
sqlalchemy
